Basic services (#35)
* Swagger config

* Convert BigInteger to Long

* Added repository, services and controllers for Pins, Comments and more

* Small fixups

+++ server/src/main/java/hr/tvz/napredna/java/dijezetserver/repository/StationRepository.java
@@ -3,7 +3,5 @@ package hr.tvz.napredna.java.dijezetserver.repository;
 import hr.tvz.napredna.java.dijezetserver.model.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface StationRepository extends JpaRepository<Station, BigInteger> {
+public interface StationRepository extends JpaRepository<Station, Long> {
 }
