+++ server/src/main/java/hr/tvz/napredna/java/dijezetserver/service/StationService.java
@@ -2,12 +2,14 @@ package hr.tvz.napredna.java.dijezetserver.service;
 
 import hr.tvz.napredna.java.dijezetserver.dto.StationDto;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface StationService {
     List<StationDto> findAll();
+
     StationDto save(StationDto stationDto);
-    StationDto update(BigInteger id, StationDto stationDto);
-    void deleteById(BigInteger id);
+
+    StationDto update(Long id, StationDto stationDto);
+
+    void deleteById(Long id);
 }
